Refactor compute workspace dependencies logic into separate class
